Fixed pattern 9 to perform well
