+++ Packages/vcs/Lib/vcsvtk/fillareautils.py
@@ -1,4 +1,5 @@
 import vtk
+from patterns import pattern_list
 
 # number of pixels per individual pattern block
 NUM_PIXELS = 16
@@ -104,306 +105,5 @@ def create_pattern(width, height, fillareastyle=None,
         fillareaopacity = 255
 
     # Create a pattern source image of the given size
-    pattern = patterns[fillareaindex](width, height, fillareacolors, fillareastyle, fillareaopacity)
+    pattern = pattern_list[fillareaindex](width, height, fillareacolors, fillareastyle, fillareaopacity)
     return pattern.render()
-
-
-class Pattern(object):
-    def __init__(self, width, height, colors, style, opacity):
-        self.width = width
-        self.height = height
-        self.colors = colors
-        self.style = style
-        self.opacity = opacity
-
-    def render(self):
-        """
-        Returns vtkImageData for pattern
-        """
-        patternSource = vtk.vtkImageCanvasSource2D()
-        patternSource.SetScalarTypeToUnsignedChar()
-        patternSource.SetExtent(0, self.width, 0, self.height, 0, 0)
-        patternSource.SetNumberOfScalarComponents(4)
-        patternSource.SetDrawColor(255, 255, 255, 0)
-        patternSource.FillBox(0, self.width, 0, self.height)
-        if self.style == 'hatch':
-            patternSource.SetDrawColor(self.colors[0],
-                                       self.colors[1],
-                                       self.colors[2],
-                                       self.opacity)
-        else:
-            patternSource.SetDrawColor(0, 0, 0, 255)
-
-        self.draw(patternSource)
-        patternSource.Update()
-        return patternSource.GetOutput()
-
-    def draw(self, pattern):
-        raise NotImplementedError("draw() not implemented for %s" % str(type(self)))
-
-
-class BottomLeftTri(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillTriangle(x, y,
-                                           x + NUM_PIXELS * 3 / 4, y,
-                                           x, y + NUM_PIXELS * 3 / 4)
-
-
-class TopRightTri(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillTriangle(x + NUM_PIXELS, y + NUM_PIXELS,
-                                           x + NUM_PIXELS / 4, y + NUM_PIXELS,
-                                           x + NUM_PIXELS, y + NUM_PIXELS * 1 / 4)
-
-
-class SmallRectDot(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x + NUM_PIXELS * 3 / 8, x + NUM_PIXELS * 5 / 8,
-                                      y + NUM_PIXELS * 3 / 8, y + NUM_PIXELS * 5 / 8)
-
-
-class CheckerBoard(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x, x + NUM_PIXELS / 2,
-                                      y, y + NUM_PIXELS / 2)
-                patternSource.FillBox(x + NUM_PIXELS / 2, x + NUM_PIXELS,
-                                      y + NUM_PIXELS / 2, y + NUM_PIXELS)
-
-
-class HorizStripe(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, self.height, NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillBox(0, self.width, lev + NUM_PIXELS / 4, lev + NUM_PIXELS * 3 / 4)
-
-
-class VertStripe(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, self.width, NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillBox(lev + NUM_PIXELS / 4, lev + NUM_PIXELS * 3 / 4, 0, self.height)
-
-
-class HorizDash(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x, x + NUM_PIXELS * 3 / 4,
-                                      y, y + NUM_PIXELS / 4)
-
-
-class VertDash(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x + NUM_PIXELS / 4, x + NUM_PIXELS / 2,
-                                      y, y + NUM_PIXELS * 3 / 4)
-                patternSource.FillBox(x + NUM_PIXELS * 3 / 4, x + NUM_PIXELS,
-                                      y, y + NUM_PIXELS * 3 / 4)
-
-
-class XDash(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillTube(x + NUM_PIXELS / 8, y + NUM_PIXELS / 8,
-                                       x + NUM_PIXELS * 3 / 8, y + NUM_PIXELS * 3 / 8,
-                                       NUM_PIXELS / 8)
-                patternSource.FillTube(x + NUM_PIXELS * 5 / 8, y + NUM_PIXELS * 5 / 8,
-                                       x + NUM_PIXELS * 7 / 8, y + NUM_PIXELS * 7 / 8,
-                                       NUM_PIXELS / 8)
-                patternSource.FillTube(x + NUM_PIXELS / 8, y + NUM_PIXELS * 7 / 8,
-                                       x + NUM_PIXELS * 3 / 8, y + NUM_PIXELS * 5 / 8,
-                                       NUM_PIXELS / 8)
-                patternSource.FillTube(x + NUM_PIXELS * 5 / 8, y + NUM_PIXELS * 3 / 8,
-                                       x + NUM_PIXELS * 7 / 8, y + NUM_PIXELS / 8,
-                                       NUM_PIXELS / 8)
-
-
-class ThinDiagDownRight(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, max(self.width, self.height) + min(self.width, self.height), NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillTube(0, lev, lev, 0, NUM_PIXELS / 8)
-
-
-class ThickDiagRownRight(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, max(self.width, self.height) + min(self.width, self.height), NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillTube(0, lev, lev, 0, NUM_PIXELS / 4)
-
-
-class ThinDiagUpRight(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, max(self.width, self.height) + min(self.width, self.height), NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillTube(lev, self.height, 0, self.height - lev, NUM_PIXELS / 8)
-
-
-class ThickDiagUpRight(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, max(self.width, self.height) + min(self.width, self.height), NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillTube(lev, self.height, 0, self.height - lev, NUM_PIXELS / 4)
-
-
-class ThickThinVertStripe(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, self.width, NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillBox(lev + NUM_PIXELS / 8, lev + NUM_PIXELS * 1 / 2, 0, self.height)
-            patternSource.FillBox(lev + NUM_PIXELS * 3 / 4, lev + NUM_PIXELS * 7 / 8, 0, self.height)
-
-
-class ThickThinHorizStripe(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        patternLevels = range(0, self.height, NUM_PIXELS)
-        for lev in patternLevels:
-            patternSource.FillBox(0, self.width, lev + NUM_PIXELS / 8, lev + NUM_PIXELS * 1 / 2)
-            patternSource.FillBox(0, self.width, lev + NUM_PIXELS * 3 / 4, lev + NUM_PIXELS * 7 / 8)
-
-
-class LargeRectDot(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x + NUM_PIXELS * 1 / 8, x + NUM_PIXELS * 7 / 8,
-                                      y + NUM_PIXELS * 1 / 8, y + NUM_PIXELS * 7 / 8)
-
-
-class Diamond(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillTriangle(x, y,
-                                           x + NUM_PIXELS / 4, y,
-                                           x, y + NUM_PIXELS / 4)
-                patternSource.FillTriangle(x, y + NUM_PIXELS,
-                                           x + NUM_PIXELS / 4, y + NUM_PIXELS,
-                                           x, y + NUM_PIXELS * 3 / 4)
-                patternSource.FillTriangle(x + NUM_PIXELS, y + NUM_PIXELS,
-                                           x + NUM_PIXELS * 3 / 4, y + NUM_PIXELS,
-                                           x + NUM_PIXELS, y + NUM_PIXELS * 3 / 4)
-                patternSource.FillTriangle(x + NUM_PIXELS, y,
-                                           x + NUM_PIXELS * 3 / 4, y,
-                                           x + NUM_PIXELS, y + NUM_PIXELS / 4)
-                patternSource.FillTriangle(x, y + NUM_PIXELS / 2,
-                                           x + NUM_PIXELS / 2, y + NUM_PIXELS,
-                                           x + NUM_PIXELS / 2, y)
-                patternSource.FillTriangle(x + NUM_PIXELS, y + NUM_PIXELS / 2,
-                                           x + NUM_PIXELS / 2, y + NUM_PIXELS,
-                                           x + NUM_PIXELS / 2, y)
-
-
-class AbstractArt(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x, x + NUM_PIXELS / 8,
-                                      y, y + NUM_PIXELS / 8)
-                patternSource.FillBox(x + NUM_PIXELS / 2, x + NUM_PIXELS * 5 / 8,
-                                      y, y + NUM_PIXELS / 8)
-                patternSource.FillBox(x, x + NUM_PIXELS / 8,
-                                      y + NUM_PIXELS / 2, y + NUM_PIXELS * 5 / 8)
-                patternSource.FillBox(x + NUM_PIXELS / 8, x + NUM_PIXELS / 2,
-                                      y + NUM_PIXELS / 8, y + NUM_PIXELS / 2)
-                patternSource.FillBox(x + NUM_PIXELS * 5 / 8, x + NUM_PIXELS,
-                                      y + NUM_PIXELS * 5 / 8, y + NUM_PIXELS)
-                patternSource.FillBox(x + NUM_PIXELS / 8, x + NUM_PIXELS * 3 / 8,
-                                      y + NUM_PIXELS * 5 / 8, y + NUM_PIXELS * 3/4)
-                patternSource.FillBox(x + NUM_PIXELS * 5 / 8, x + NUM_PIXELS * 7 / 8,
-                                      y + NUM_PIXELS / 8, y + NUM_PIXELS * 1 / 4)
-
-
-class Snake(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.FillBox(x, x + NUM_PIXELS * 3 / 8,
-                                      y, y + NUM_PIXELS / 4)
-                patternSource.FillBox(x + NUM_PIXELS * 5 / 8, x + NUM_PIXELS,
-                                      y, y + NUM_PIXELS / 4)
-                patternSource.FillBox(x + NUM_PIXELS / 4, x + NUM_PIXELS * 3 / 4,
-                                      y + NUM_PIXELS / 3, y + NUM_PIXELS * 2 / 3)
-
-
-class EmptyCircle(Pattern):
-    def draw(self, patternSource):
-        if patternSource is None:
-            return None
-        global NUM_PIXELS
-        for x in xrange(0, self.width, NUM_PIXELS):
-            for y in xrange(0, self.height, NUM_PIXELS):
-                patternSource.DrawCircle(x + NUM_PIXELS / 2, y + NUM_PIXELS / 2, NUM_PIXELS / 4.)
-
-
-patterns = [Pattern, BottomLeftTri, TopRightTri, SmallRectDot, CheckerBoard,
-            HorizStripe, VertStripe, HorizDash, VertDash, XDash, ThinDiagDownRight,
-            ThickDiagRownRight, ThinDiagUpRight, ThickDiagUpRight, ThickThinVertStripe,
-            ThickThinHorizStripe, LargeRectDot, Diamond, AbstractArt, Snake, EmptyCircle]
